+++ src/Aevatar.Application.Contracts/ApiKeys/ApiKeyListResponseDto.cs
@@ -5,6 +5,7 @@ namespace Aevatar.ApiKeys;
 
 public class ApiKeyListResponseDto
 {
+    public Guid Id { get; set; }
     public Guid ProjectId { get; set; }
     public string ApiKeyName { get; set; }
     public string ApiKey { get; set; }
