+++ src/Aevatar.Application/Service/ProjectApiKeyService.cs
@@ -12,14 +12,6 @@ using Volo.Abp.Identity;
 
 namespace Aevatar.Service;
 
-public interface IProjectApiKeyService
-{
-    Task CreateAsync(Guid projectId, string keyName);
-    Task DeleteAsync(Guid apiKeyId);
-    Task ModifyApiKeyAsync(Guid apiKeyId, string keyName);
-    Task<List<ApiKeyListResponseDto>> GetApiKeysAsync(Guid projectId);
-}
-
 public class ProjectApiKeyService : ApplicationService, IProjectApiKeyService
 {
     private readonly IApiKeysRepository _apiKeysRepository;
@@ -36,11 +28,8 @@ public class ProjectApiKeyService : ApplicationService, IProjectApiKeyService
         _identityUserManager = identityUserManager;
     }
 
-
     public async Task CreateAsync(Guid projectId, string keyName)
     {
-        // todo:validate create rights
-
         _logger.LogDebug($"[ProjectApiKeyService][CreateAsync] projectId:{projectId}, keyName:{keyName}");
 
         if (await _apiKeysRepository.CheckProjectApiKeyNameExist(projectId, keyName))
@@ -48,17 +37,9 @@ public class ProjectApiKeyService : ApplicationService, IProjectApiKeyService
             throw new BusinessException(message: "key name has exist");
         }
 
-        var random = new Random();
-        var randNum = random.Next(0, 1000000000);
-        var apikeyStr = MD5Util.CalculateMD5($"{projectId.ToString()}-{keyName}-{randNum}");
+        var apikeyStr = MD5Util.CalculateMD5($"{projectId.ToString()}-{keyName}-{Guid.NewGuid()}");
+        var apiKey = new ApiKeyInfo(Guid.NewGuid(), projectId, keyName, apikeyStr);
 
-        var creator = _appService.GetCurrentUserId();
-        var apiKey = new ApiKeyInfo(Guid.NewGuid(), projectId, keyName, apikeyStr)
-        {
-            CreationTime = DateTime.Now,
-            CreatorId = creator,
-        };
-        
         await _apiKeysRepository.InsertAsync(apiKey);
     }
 
@@ -73,7 +54,7 @@ public class ProjectApiKeyService : ApplicationService, IProjectApiKeyService
     {
         // todo:validate modify rights
         _logger.LogDebug($"[ProjectApiKeyService][ModifyApiKeyAsync] apiKeyId:{apiKeyId}, keyName:{keyName}");
-        
+
         var apiKeyInfo = await _apiKeysRepository.GetAsync(apiKeyId);
         if (apiKeyInfo == null)
         {
@@ -84,12 +65,11 @@ public class ProjectApiKeyService : ApplicationService, IProjectApiKeyService
         {
             throw new BusinessException(message: "key name has exist");
         }
-        
+
         if (apiKeyInfo.ApiKeyName == keyName)
         {
             throw new BusinessException(message: "ApiKey is the same ");
         }
-        
 
         apiKeyInfo.ApiKeyName = keyName;
 
@@ -99,15 +79,16 @@ public class ProjectApiKeyService : ApplicationService, IProjectApiKeyService
     public async Task<List<ApiKeyListResponseDto>> GetApiKeysAsync(Guid projectId)
     {
         // todo:validate GetApiKeysAsync rights
-        
+
         var apiKeyList = await _apiKeysRepository.GetProjectApiKeys(projectId, 10, 0);
         var result = new List<ApiKeyListResponseDto>();
         foreach (var item in apiKeyList)
         {
             var creatorInfo = await _identityUserManager.GetByIdAsync((Guid)item.CreatorId!);
-            
+
             result.Add(new ApiKeyListResponseDto()
             {
+                Id = item.Id,
                 ApiKey = item.ApiKey,
                 ApiKeyName = item.ApiKeyName,
                 CreateTime = item.CreationTime,
