--- src/Aevatar.MongoDB/ApiKeys/ApiKeysMongoDbContext.cs
@@ -1,10 +1,0 @@
-using Aevatar.ApiKey;
-using MongoDB.Driver;
-using Volo.Abp.MongoDB;
-
-namespace Aevatar.ApiKeys;
-
-public class ApiKeysMongoDbContext : AbpMongoDbContext
-{
-    public IMongoCollection<ApiKeyInfo> ApiKeyInfoCollection => Collection<ApiKeyInfo>();
-}
