+++ src/Aevatar.MongoDB/MongoDb/AevatarMongoDbModule.cs
@@ -36,10 +36,6 @@ public class AevatarMongoDbModule : AbpModule
         });
 
         context.Services.AddTransient<IApiKeysRepository, ApiKeyMongoRepository>();
-        context.Services.AddMongoDbContext<ApiKeysMongoDbContext>(options =>
-        {
-            options.AddDefaultRepositories();
-        });
 
         Configure<AbpUnitOfWorkDefaultOptions>(options =>
         {
